+++ resources/lang/de/companies.php
@@ -1,0 +1,7 @@
+<?php
+
+return [
+    'unnamed' => 'Firma ohne Namen',
+    'title' => 'Firmen',
+    'singular' => 'Firma',
+];